04/12 spring 과제 완료

+++ src/main/java/ezenweb/web/service/MemberService.java
@@ -4,35 +4,155 @@ import ezenweb.web.domain.member.MemberDto;
 import ezenweb.web.domain.member.MemberEntity;
 import ezenweb.web.domain.member.MemberEntityRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
+import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
+import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
+import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
+import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
-import java.util.Optional;
+import java.util.*;
+
+// UserDetailsService : 일반유저 서비스 구현 ---> loadUserByUsername 구현
+// OAuth2UserService<OAuth2UserRequest , OAuth2User>  : oauth2 유저 서비스 구현 ---> loadUser 구현
 
 @Service // 서비스 레이어
 @Slf4j
-public class MemberService implements UserDetailsService {
+public class MemberService implements UserDetailsService , OAuth2UserService<OAuth2UserRequest , OAuth2User> {
+
+    @Override // 토큰 결과 [ JSON { 필드명 : 값 , 필드명 : 값  } VS Map { 키 = 값 , 키 = 값 , 키 = 값 } ]
+    public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
+        // 1. 인증[로그인] 결과 토큰 확인
+        log.info( "토큰 결과 정보 : " +userRequest );
+        // 2. 전달받은 토큰을 이용한 회원정보 요청
+        OAuth2UserService oAuth2UserService = new DefaultOAuth2UserService();
+        log.info( "서비스 정보 : " +oAuth2UserService.loadUser( userRequest ) );
+        OAuth2User oAuth2User = oAuth2UserService.loadUser( userRequest );
+        log.info("회원정보 : " + oAuth2User.getAuthorities() );
+        // !!!! : oAuth2User.getAttributes()  map< String , Object >구조
+        //  {sub=114044778334166488538, name=아이티단자, given_name=단자,email=kgs2072@naver.com}
+
+        // 3. 클라이언트 ID 식별해서 처리 (구글/카카오등등..)
+        String registrationId = userRequest.getClientRegistration().getRegistrationId();
+        log.info("클라이언트 ID : " + registrationId );
+
+        String email = null;
+        String name = null;
+        if(registrationId.equals("kakao")){
+           Map<String , Object> kakao_account = (Map<String , Object>)oAuth2User.getAttributes().get("kakao_account");
+            Map<String , Object> profile = (Map<String , Object>)kakao_account.get("profile");
+           log.info(" kakao_account : " + kakao_account );
+           log.info(" profile : " + profile );
+            email = (String)kakao_account.get("email");
+            name = (String)profile.get("nickname");
+            log.info(" email : " + email );
+            log.info(" name : " + name );
+        }
+        else if(registrationId.equals("naver")){
+            Map<String, Object> response = (Map<String, Object>)oAuth2User.getAttributes().get("response");
+
+            log.info(" response : " + response );
+
+            email = (String)response.get("email");
+            name = (String)response.get("nickname");
+            log.info(" email : " + email );
+            log.info(" name : " + name );
+
+        }
+        else if(registrationId.equals("google")){
+            // 구글의 이메일 호출
+            email =  (String)oAuth2User.getAttributes().get( "email" );  log.info(" google name : " + email );
+            // 구글의 이름 호출
+            name =  (String)oAuth2User.getAttributes().get( "name" );    log.info(" google email : " + name );
+        }
 
 
+
+
+        // 인가 객체 [ OAuth2User----> MemberDto 통합Dto( 일반+oauth) ]
+        MemberDto memberDto = new MemberDto();
+        memberDto.set소셜회원정보( oAuth2User.getAttributes() );
+        memberDto.setMemail( email );
+        memberDto.setMname( name );
+        Set<GrantedAuthority> 권한목록 = new HashSet<>();
+        SimpleGrantedAuthority 권한 = new SimpleGrantedAuthority("ROLE_user");
+        권한목록.add( 권한 );
+        memberDto.set권한목록( 권한목록 );
+        // 1. DB 저장하기 전에 해당 이메일로 된 이메일 존재하는지 검사
+        MemberEntity entity = memberEntityRepository.findByMemail( email );
+        if( entity == null ){ // 첫방문
+            // DB 처리 [ 첫 방문시에만 db등록  , 두번째 방문시 부터는 db수정  ]
+            memberDto.setMrole("oauthuser"); // DB에 저장할 권한명
+            memberEntityRepository.save( memberDto.toEntity() );
+        }else{// 두번째 방문 이상 수정 처리
+            entity.setMname( name );
+        }
+        return memberDto;
+    }
 
     @Autowired
     private MemberEntityRepository memberEntityRepository;
     @Autowired private HttpServletRequest request; // 요청 객체
 
-    // 1. 회원가입
+
+    // 아아디 찾기
+
+    @Transactional
+    public String findId(String mname , String mphone){
+        MemberEntity findId = memberEntityRepository.findByMnameAndMphone( mname , mphone).get();
+
+        log.info("아이디 찾기 : " + findId.getMemail());
+        MemberDto dto = findId.todto();
+
+        return dto.getMemail();
+    }
+
+    // 비밀번호 찾기
+    @Transactional
+    public String findPw(String memail, String mphone){
+        Optional<MemberEntity> entityOptional = memberEntityRepository.findByMemailAndMphone( memail, mphone);
+        log.info("찾은 맴버 : " + entityOptional);
+        Random random = new Random();
+        int new_pw = random.nextInt( 99999999 );
+
+        String npw = Integer.toString( new_pw );
+        BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+        MemberEntity entity = entityOptional.get();
+        entity.setMpassword( passwordEncoder.encode(npw));
+
+
+        MemberDto dto = entity.todto();
+
+        memberEntityRepository.save( dto.toEntity() );
+
+        log.info("비밀번호 찾기 : " + npw);
+
+        return npw;
+
+    }
+    // 1. 일반 회원가입 [ 본 애플리케이션에서 가입한 회원  ]
     @Transactional
     public boolean write( MemberDto memberDto ){
+        MemberEntity findId = memberEntityRepository.findByMemail(memberDto.getMemail());
+        log.info("회원가입시 아이디 중복체크 : " + findId );
+        if(findId != null){
+            return false;
+        }
+
+
+
         // 스프링 시큐리티에서 제공하는 암호화[ 사람이 이해하기 어렵고 컴퓨터는 이해할수 있는 단어 ] 사용하기
         // DB 내에서 패스워드 감추기 , 정보가 이동하면 패스워드 노출 방지
         BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
@@ -40,6 +160,9 @@ public class MemberService implements UserDetailsService {
         log.info("비크립트 암호화 사용 : " + passwordEncoder.encode( "1234") );
         // 입력받은[DTO] 패스워드 암호화 해서 다시 DTO에 저장하자.
         memberDto.setMpassword( passwordEncoder.encode( memberDto.getMpassword() ) );
+
+        // 등급/권한 부여
+        memberDto.setMrole("user");
 
         MemberEntity entity = memberEntityRepository.save( memberDto.toEntity() );
         if( entity.getMno() > 0 ){ return  true;}
@@ -84,22 +207,28 @@ public class MemberService implements UserDetailsService {
     @Transactional
     public boolean update(  MemberDto memberDto ){
         Optional<MemberEntity> entityOptional = memberEntityRepository.findById( memberDto.getMno() );
+
         if( entityOptional.isPresent() ){
             MemberEntity entity = entityOptional.get();
             entity.setMname( memberDto.getMname());  entity.setMphone( memberDto.getMphone() );
-            entity.setMrole( memberDto.getMrole() ); entity.setMpassword( memberDto.getMpassword());
             return true;
         }
         return false;
     }
     // 4. 회원탈퇴
     @Transactional
-    public boolean delete( int mno ){
+    public boolean delete( int mno , String mpassword ){
         Optional<MemberEntity> entityOptional = memberEntityRepository.findById( mno );
+        BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+
         if( entityOptional.isPresent()){
             MemberEntity entity = entityOptional.get();
-            memberEntityRepository.delete( entity );
-            return true;
+            if(passwordEncoder.matches( mpassword , entity.getMpassword() )){
+                memberEntityRepository.delete( entity );
+                return true;
+            }
+
+
         }
         return false;
     }
@@ -113,7 +242,6 @@ public class MemberService implements UserDetailsService {
     // [ 스프링 시큐리티 적용했을때 사용되는 로그인 메소드 ]
     @Override
     public UserDetails loadUserByUsername(String memail ) throws UsernameNotFoundException {
-       log.info( "memail : " + memail );
         // 1. UserDetailsService 인터페이스 구현
         // 2. loadUserByUsername() 메소드 : 아이디 검증
         // 패스워드 검증 [ 시큐리티 자동 ]
@@ -121,8 +249,19 @@ public class MemberService implements UserDetailsService {
         if( entity == null ){ return null; }
         // 3. 검증후 세션에 저장할 DTO 반환
         MemberDto dto = entity.todto();
+        // Dto 권한(여러개) 넣어주기
+        // 1. 권한목록 만들기
+        Set<GrantedAuthority> 권한목록 = new HashSet<>();
+        // 2. 권한객체 만들기 [ DB 존재하는 권한명( ROLE_!!!! )으로  ]
+        // 권한 없을경우 : ROLE_ANONYMOUS  / 권한 있을경우 ROLE_권한명 : ROLE_admin , ROLE_user
+        SimpleGrantedAuthority 권한명1 = new SimpleGrantedAuthority( "ROLE_"+entity.getMrole() );
+        // 3. 만든 권한객체를 권한목록[컬렉션]에  추가
+        권한목록.add( 권한명1 );
+        // 4. UserDetails 에 권한목록 대입
+        dto.set권한목록( 권한목록 );
+
         log.info( "dto:" + dto);
-        return dto;
+        return dto; // UserDetails : 원본데이터의 검증할 계정 , 패스워드 포함
     }
     // 2. [세션에 존재하는 ] 회원정보
     @Transactional
@@ -131,14 +270,13 @@ public class MemberService implements UserDetailsService {
         // SecurityContextHolder : 시큐리티 정보 저장소
         // SecurityContextHolder.getContext()  : 시큐리티 저장된 정보 호출
         // SecurityContextHolder.getContext().getAuthentication();  // 인증 전체 정보 호출
-        // log.info(" Auth : " + SecurityContextHolder.getContext().getAuthentication() );
-        // log.info(" Auth : " + SecurityContextHolder.getContext().getAuthentication().getPrincipal() );
+        log.info(" Auth : " + SecurityContextHolder.getContext().getAuthentication() );
+        log.info(" Auth : " + SecurityContextHolder.getContext().getAuthentication().getPrincipal() );
         Object o = SecurityContextHolder.getContext().getAuthentication().getPrincipal(); // 인증된 회원의 정보 호출
-        log.info(" o : " + o );
-        if( o == null ){ return null; }
-
+        if( o.equals("anonymousUser") ){ return null; }
+        // [ Principal ]  // 인증  실패시 : anonymousUser   // 인증  성공시 : 회원정보[Dto]
         // 2. 인증된 객체내 회원정보[ Principal ] 타입 변환
-        return (MemberDto) o; //  Object ---> dto
+        return (MemberDto)o; //  Object ---> dto
         /*
         // 2. 일반 세션으로 로그인 정보를 관리 했을때 [ JSP ]
         String memail = (String)request.getSession().getAttribute("login");
